+++ Items/Item.cs
@@ -8,7 +8,7 @@ namespace Core
         public abstract ItemType Type { get; }
         public abstract string Name { get; }
         public abstract string? ChemicalFormula { get; }
-        public uint Quantity { get; private set; }
+        public ulong Quantity { get; private set; }
         public virtual UnitType Units => UnitType.Unit;
         public ulong Id { get; set; }
 
@@ -46,7 +46,7 @@ namespace Core
             this.Id = IdGenerator.GenerateId();
         }
 
-        public void AddToStack(uint amount)
+        public void AddToStack(ulong amount)
         {
             if (Quantity + amount > MaxStack)
                 throw new InvalidOperationException("Cannot add to stack, would exceed max stack size.");
@@ -54,7 +54,7 @@ namespace Core
             Quantity += amount;
         }
 
-        public void RemoveFromStack(uint amount)
+        public void RemoveFromStack(ulong amount)
         {
             if (amount > Quantity)
                 throw new InvalidOperationException("Cannot remove from stack, would go below 0.");
@@ -62,7 +62,7 @@ namespace Core
             Quantity -= amount;
         }
 
-        public void SetQuantity(uint quantity)
+        public void SetQuantity(ulong quantity)
         {
             if (quantity > MaxStack)
                 throw new InvalidOperationException("Cannot set quantity, would exceed max stack size.");
